requirements.txt
requests
ruamel.yaml
alpinepkgs